+++ certbot-dns-cloudflare/tests/dns_cloudflare_test.py
@@ -41,7 +41,7 @@ class AuthenticatorTest(test_util.TempDirTestCase, dns_test_common.BaseAuthentic
     def test_perform(self):
         self.auth.perform([self.achall])
 
-        expected = [mock.call.add_txt_record(DOMAIN, '_acme-challenge.'+DOMAIN, mock.ANY, mock.ANY)]
+        expected = [mock.call.add_txt_record('_acme-challenge.'+DOMAIN, mock.ANY, mock.ANY)]
         self.assertEqual(expected, self.mock_client.mock_calls)
 
     def test_cleanup(self):
@@ -49,7 +49,7 @@ class AuthenticatorTest(test_util.TempDirTestCase, dns_test_common.BaseAuthentic
         self.auth._attempt_cleanup = True
         self.auth.cleanup([self.achall])
 
-        expected = [mock.call.del_txt_record(DOMAIN, '_acme-challenge.'+DOMAIN, mock.ANY)]
+        expected = [mock.call.del_txt_record('_acme-challenge.'+DOMAIN, mock.ANY)]
         self.assertEqual(expected, self.mock_client.mock_calls)
 
     def test_api_token(self):
@@ -57,7 +57,7 @@ class AuthenticatorTest(test_util.TempDirTestCase, dns_test_common.BaseAuthentic
                               self.config.cloudflare_credentials)
         self.auth.perform([self.achall])
 
-        expected = [mock.call.add_txt_record(DOMAIN, '_acme-challenge.'+DOMAIN, mock.ANY, mock.ANY)]
+        expected = [mock.call.add_txt_record('_acme-challenge.'+DOMAIN, mock.ANY, mock.ANY)]
         self.assertEqual(expected, self.mock_client.mock_calls)
 
     def test_no_creds(self):
@@ -115,7 +115,7 @@ class CloudflareClientTest(unittest.TestCase):
     def test_add_txt_record(self):
         self.cf.zones.get.return_value = [{'id': self.zone_id}]
 
-        self.cloudflare_client.add_txt_record(DOMAIN, self.record_name, self.record_content,
+        self.cloudflare_client.add_txt_record(self.record_name, self.record_content,
                                               self.record_ttl)
 
         self.cf.zones.dns_records.post.assert_called_with(self.zone_id, data=mock.ANY)
@@ -135,7 +135,7 @@ class CloudflareClientTest(unittest.TestCase):
         self.assertRaises(
             errors.PluginError,
             self.cloudflare_client.add_txt_record,
-            DOMAIN, self.record_name, self.record_content, self.record_ttl)
+            self.record_name, self.record_content, self.record_ttl)
 
     def test_add_txt_record_error_during_zone_lookup(self):
         self.cf.zones.get.side_effect = API_ERROR
@@ -143,7 +143,7 @@ class CloudflareClientTest(unittest.TestCase):
         self.assertRaises(
             errors.PluginError,
             self.cloudflare_client.add_txt_record,
-            DOMAIN, self.record_name, self.record_content, self.record_ttl)
+            self.record_name, self.record_content, self.record_ttl)
 
     def test_add_txt_record_zone_not_found(self):
         self.cf.zones.get.return_value = []
@@ -151,32 +151,32 @@ class CloudflareClientTest(unittest.TestCase):
         self.assertRaises(
             errors.PluginError,
             self.cloudflare_client.add_txt_record,
-            DOMAIN, self.record_name, self.record_content, self.record_ttl)
+            self.record_name, self.record_content, self.record_ttl)
 
     def test_add_txt_record_bad_creds(self):
         self.cf.zones.get.side_effect = CloudFlare.exceptions.CloudFlareAPIError(6003, '', '')
         self.assertRaises(
             errors.PluginError,
             self.cloudflare_client.add_txt_record,
-            DOMAIN, self.record_name, self.record_content, self.record_ttl)
+            self.record_name, self.record_content, self.record_ttl)
 
         self.cf.zones.get.side_effect = CloudFlare.exceptions.CloudFlareAPIError(9103, '', '')
         self.assertRaises(
             errors.PluginError,
             self.cloudflare_client.add_txt_record,
-            DOMAIN, self.record_name, self.record_content, self.record_ttl)
+            self.record_name, self.record_content, self.record_ttl)
 
         self.cf.zones.get.side_effect = CloudFlare.exceptions.CloudFlareAPIError(9109, '', '')
         self.assertRaises(
             errors.PluginError,
             self.cloudflare_client.add_txt_record,
-            DOMAIN, self.record_name, self.record_content, self.record_ttl)
+            self.record_name, self.record_content, self.record_ttl)
 
     def test_del_txt_record(self):
         self.cf.zones.get.return_value = [{'id': self.zone_id}]
         self.cf.zones.dns_records.get.return_value = [{'id': self.record_id}]
 
-        self.cloudflare_client.del_txt_record(DOMAIN, self.record_name, self.record_content)
+        self.cloudflare_client.del_txt_record(self.record_name, self.record_content)
 
         expected = [mock.call.zones.get(params=mock.ANY),
                     mock.call.zones.dns_records.get(self.zone_id, params=mock.ANY),
@@ -193,14 +193,14 @@ class CloudflareClientTest(unittest.TestCase):
     def test_del_txt_record_error_during_zone_lookup(self):
         self.cf.zones.get.side_effect = API_ERROR
 
-        self.cloudflare_client.del_txt_record(DOMAIN, self.record_name, self.record_content)
+        self.cloudflare_client.del_txt_record(self.record_name, self.record_content)
 
     def test_del_txt_record_error_during_delete(self):
         self.cf.zones.get.return_value = [{'id': self.zone_id}]
         self.cf.zones.dns_records.get.return_value = [{'id': self.record_id}]
         self.cf.zones.dns_records.delete.side_effect = API_ERROR
 
-        self.cloudflare_client.del_txt_record(DOMAIN, self.record_name, self.record_content)
+        self.cloudflare_client.del_txt_record(self.record_name, self.record_content)
         expected = [mock.call.zones.get(params=mock.ANY),
                     mock.call.zones.dns_records.get(self.zone_id, params=mock.ANY),
                     mock.call.zones.dns_records.delete(self.zone_id, self.record_id)]
@@ -211,7 +211,7 @@ class CloudflareClientTest(unittest.TestCase):
         self.cf.zones.get.return_value = [{'id': self.zone_id}]
         self.cf.zones.dns_records.get.side_effect = API_ERROR
 
-        self.cloudflare_client.del_txt_record(DOMAIN, self.record_name, self.record_content)
+        self.cloudflare_client.del_txt_record(self.record_name, self.record_content)
         expected = [mock.call.zones.get(params=mock.ANY),
                     mock.call.zones.dns_records.get(self.zone_id, params=mock.ANY)]
 
@@ -221,7 +221,7 @@ class CloudflareClientTest(unittest.TestCase):
         self.cf.zones.get.return_value = [{'id': self.zone_id}]
         self.cf.zones.dns_records.get.return_value = []
 
-        self.cloudflare_client.del_txt_record(DOMAIN, self.record_name, self.record_content)
+        self.cloudflare_client.del_txt_record(self.record_name, self.record_content)
         expected = [mock.call.zones.get(params=mock.ANY),
                     mock.call.zones.dns_records.get(self.zone_id, params=mock.ANY)]
 
@@ -230,7 +230,7 @@ class CloudflareClientTest(unittest.TestCase):
     def test_del_txt_record_no_zone(self):
         self.cf.zones.get.return_value = [{'id': None}]
 
-        self.cloudflare_client.del_txt_record(DOMAIN, self.record_name, self.record_content)
+        self.cloudflare_client.del_txt_record(self.record_name, self.record_content)
         expected = [mock.call.zones.get(params=mock.ANY)]
 
         self.assertEqual(expected, self.cf.mock_calls)