+++ certbot/certbot/plugins/dns_common.py
@@ -87,7 +87,7 @@ class DNSAuthenticator(common.Plugin):
         """
         Performs a dns-01 challenge by creating a DNS TXT record.
 
-        :param str domain: The domain being validated.
+        :param str domain: The domain being validated. For logging purposes only.
         :param str validation_domain_name: The validation record domain name.
         :param str validation: The validation record content.
         :raises errors.PluginError: If the challenge cannot be performed
@@ -101,7 +101,7 @@ class DNSAuthenticator(common.Plugin):
 
         Fails gracefully if no such record exists.
 
-        :param str domain: The domain being validated.
+        :param str domain: The domain being validated. For logging purposes only.
         :param str validation_domain_name: The validation record domain name.
         :param str validation: The validation record content.
         """